libraries/workflow/plugins/python/requirements.txt
# MetaBuilder Python Workflow Plugins - Master Requirements File
# This file consolidates all dependencies across Python workflow plugins
# Last Updated: 2026-01-23
# Python Runtime: 3.9+

# Core Dependencies - Required by all plugins
python-dotenv>=1.0.0  # Backend: load_env plugin
tenacity>=8.2.3       # Core: exponential backoff for API calls

# API Clients - Backend plugins
openai>=1.3.0         # backend_create_openai
slack-sdk>=3.23.0     # backend_create_slack, notifications_slack
PyGithub>=2.1.1       # backend_create_github
discord.py>=2.3.2     # backend_create_discord, notifications_discord

# Web Framework
Flask>=3.0.0          # web_* plugins for creating Flask apps
Flask-CORS>=4.0.0     # web_* plugins for CORS support (implicit from Flask dependencies)

# Web/HTTP Utilities
requests>=2.31.0      # tools_create_pull_request, web utilities

# Data Processing & Validation
pydantic>=2.5.0       # packagerepo: normalize_entity, validate_entity

# Database (if packagerepo plugins need DB access)
PyJWT>=2.8.1          # packagerepo: auth_verify_jwt

# Testing Frameworks
pytest>=7.4.3         # test_* plugins
pytest-asyncio>=0.21.1  # For async test execution

# Optional/Conditional Dependencies (installed on demand)
# Note: These are dynamically imported in plugins, so install as needed

# For advanced logging (optional)
# logging-json>=2.0.0

# For database support (optional, uncomment if needed)
# psycopg2-binary>=2.9.9  # PostgreSQL
# pymongo>=4.6.0          # MongoDB
# redis>=5.0.1            # Redis

# For distributed task processing (optional)
# celery>=5.3.4
# kafka-python>=2.0.2
